+++ drylab/templates/drylab/menu.html
@@ -68,16 +68,6 @@
                         </ul>
                     </li>
                     {% endif %}
-                    {% if user.username == "admin" %}
-                    <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" data-bs-toggle="dropdown"
-                        aria-expanded="false">USER ACTIVITY</a>
-                        <ul class="dropdown-menu">
-                            <li><a class="dropdown-item" href="/drylab/openSessions">CURRENT OPEN USER SESSIONS</a></li>
-                            <li><a class="dropdown-item" href="/drylab/userLogin">USER LAST LOGING</a></li>
-                        </ul>
-                    </li>
-                    {% endif %}
                 </ul>
             </div>
         </div>
