+++ drylab/templates/drylab/search_service.html
@@ -108,12 +108,12 @@
                                                 <label for="service_state">Service state</label>
                                             </div>
                                             <div class="form-floating mb-1">
-                                                <input type="text" class="form-control" id="start_date"
+                                                <input type="date" class="form-control" id="start_date"
                                                     name="start_date" placeholder="From date (DD-MM-YYYY)">
                                                 <label for="start_date">From date (DD-MM-YYYY)</label>
                                             </div>
                                             <div class="form-floating mb-1">
-                                                <input type="text" class="form-control" id="end_date" name="end_date"
+                                                <input type="date" class="form-control" id="end_date" name="end_date"
                                                     placeholder="From date (DD-MM-YYYY)">
                                                 <label for="end_date">To date (DD-MM-YYYY)</label>
                                             </div>
