+++ drylab/templates/drylab/stats_services_time.html
@@ -69,8 +69,6 @@
 					</div>
 				</div>
 			</div>
-		</div>
-		<div class="row justify-content-center pb-2 pt-2">
 			<div class="col-md-6 ">
 				<div class="card ">
 					<div class="card-header">
@@ -117,66 +115,66 @@
 				</div>
 			</div>
 		</div>
-	</div>
-	<div class="row justify-content-center pb-2 pt-2">
-		<div class="col-md-12 ">
-			<div class="card ">
-				<div class="card-header">
-					<h4> Level 2 available services </h4>
-				</div>
-				<div class="card-body">
-					<h4> {{services_stats_info.period_time}}.</h4>
-					{% if services_stats_info.graphic_req_l2_services %}
-					<div id="chart-7"> </div>
-					{{ services_stats_info.graphic_req_l2_services |safe }}
-					{% endif %}
-				</div>
-			</div>
-		</div>
-	</div>
-	<div class="row justify-content-center pb-2 pt-2">
-		<div class="col-md-12 ">
-			<div class="card ">
-				<div class="card-header">
-					<h4>Level 3 Available Services </h4>
-				</div>
-				<div class="card-body">
-					<h4> {{services_stats_info.period_time}}.</h4>
-					{% if services_stats_info.graphic_req_l3_services %}
-					<div id="chart-8"> </div>
-					{{ services_stats_info.graphic_req_l3_services |safe }}
-					{% endif %}
+		<div class="row justify-content-center pb-2 pt-2">
+			<div class="col-md-auto ">
+				<div class="card ">
+					<div class="card-header">
+						<h4> Level 2 available services </h4>
+					</div>
+					<div class="card-body">
+						<h4> {{services_stats_info.period_time}}.</h4>
+						{% if services_stats_info.graphic_req_l2_services %}
+						<div id="chart-7"> </div>
+						{{ services_stats_info.graphic_req_l2_services |safe }}
+						{% endif %}
+					</div>
 				</div>
 			</div>
 		</div>
-	</div>
-	{% else %}
-	<div class="row justify-content-center pb-2 pt-2">
-		<div class="col-md-8">
-			<div class="card ">
-				<div class="card-header">
-					<h3>Services statistics</h3>
+		<div class="row justify-content-center pb-2 pt-2">
+			<div class="col-md-auto">
+				<div class="card ">
+					<div class="card-header">
+						<h4>Level 3 Available Services </h4>
+					</div>
+					<div class="card-body">
+						<h4> {{services_stats_info.period_time}}.</h4>
+						{% if services_stats_info.graphic_req_l3_services %}
+						<div id="chart-8"> </div>
+						{{ services_stats_info.graphic_req_l3_services |safe }}
+						{% endif %}
+					</div>
 				</div>
-				<div class="card-body">
-					<form method="post" enctype="multipart/form-data" name="service_statistics" id="service_statistics"
-						class="form-horizontal well">
-						{% csrf_token %}
-						<input type="hidden" name="action" value="service_statistics" />
-						<div class="form-floating mb-1">
-							<input type="text" class="form-control" id="start_date"
-								name="start_date" placeholder="From date (DD-MM-YYYY)">
-							<label for="start_date">From date (DD-MM-YYYY)</label>
-						</div>
-						<div class="form-floating mb-1">
-							<input type="text" class="form-control" id="end_date" name="end_date"
-								placeholder="From date (DD-MM-YYYY)">
-							<label for="end_date">To date (DD-MM-YYYY)</label>
-						</div>
-						<input class="btn pull-left btn-outline-primary" type="reset" value="Reset Fields">
-						<input class="btn float-end btn-outline-primary" type="submit" value="Submit">
-					</form>
+			</div>
+		</div>
+		{% else %}
+		<div class="row justify-content-center pb-2 pt-2">
+			<div class="col-md-8">
+				<div class="card ">
+					<div class="card-header">
+						<h4>Services statistics</h4>
+					</div>
+					<div class="card-body">
+						<form method="post" enctype="multipart/form-data" name="service_statistics"
+							id="service_statistics" class="form-horizontal well">
+							{% csrf_token %}
+							<input type="hidden" name="action" value="service_statistics" />
+							<div class="form-floating mb-1">
+								<input type="date" class="form-control" id="start_date" name="start_date"
+									placeholder="From date (DD-MM-YYYY)">
+								<label for="start_date">From date (DD-MM-YYYY)</label>
+							</div>
+							<div class="form-floating mb-1">
+								<input type="date" class="form-control" id="end_date" name="end_date"
+									placeholder="From date (DD-MM-YYYY)">
+								<label for="end_date">To date (DD-MM-YYYY)</label>
+							</div>
+							<input class="btn pull-left btn-outline-secondary" type="reset" value="Reset Fields">
+							<input class="btn float-end btn-outline-primary" type="submit" value="Submit">
+						</form>
+						<p> Fields marked with <FONT COLOR="Red">*</FONT> are mandatory</p>
+					</div>
 				</div>
-				<p> Fields marked with <FONT COLOR="Red">*</FONT> are mandatory</p>
 			</div>
 		</div>
 	</div>
