--- drylab/templates/drylab/userLogin.html
@@ -1,37 +1,0 @@
-{% extends 'drylab/base.html' %}
-{% load static %}
-{% block content %}
-{% csrf_token %}
-    <h1><div class="color-heading">User defined on iSkyLIMS</div></h1>
-    <div class="row my-2">
-        <div class="col-sm-8">  <!-- general information from service -->
-            <div class="card ">
-                  <div class="card-header">The following tabla displays the last user login to iSkyLIMs </div>
-                  <div class="card-body">
-                      <table class="table table-hover">
-                          <thead>
-                            <tr>
-                                <th>user name</th>
-                                <th>First Name</th>
-                                <th>Last Name</th>
-                                <th>email</th>
-                                <th>last_login</th>
-                            </tr>
-                          </thead>
-                           <tbody>
-                              {% for user, fname, lname, email, login_date in login_data.user_data %}  
-                                 <tr>
-                                   <td>{{ user }}</td>
-                                   <td>{{ fname }}</td>
-                                   <td>{{ lname }}</td>
-                                   <td>{{ email }}</td>
-                                   <td>{{ login_date }}</td>
-                                 </tr>
-                              {% endfor %}   
-                           </tbody>
-                        </table>
-                    </div> <!--// end panel body -->
-                </div>
-        </div> <!--// end col-sm-8 -->
-    </div> <!--// end row -->
-{% endblock %}
